requirements.txt
torch
pytorch-cuda==12.1
einops
accelerate
torchtyping
scikit-learn
seaborn
tensorboard
wandb
polars>=0.20.19
pandas
pyarrow
marimo
gcamreader>=1.4.0
pytest
pytest-cov
click
rich
tqdm
python-dotenv
python-configuration[toml]
lxml
geopandas>=0.9.0
